Split folder for key and value schema

+++ esque/messages/avromessage.py
@@ -41,29 +41,33 @@ class AvroFileWriter(FileWriter):
             message
         )
 
-        if self.schema_changed(decoded_message) or self.schema_dir_name is None:
-            self.schema_dir_name = f"{next(self.schema_version):04}_{key_schema_id}_{value_schema_id}"
-            self.current_key_schema_id = key_schema_id
-            self.current_value_schema_id = value_schema_id
-            self._dump_schemata(key_schema_id, value_schema_id)
+        if value_schema_id != -1:
+            value_schema_dir_name = f"value_schema_{value_schema_id}"
+            self._dump_value_schema(value_schema_id, value_schema_dir_name)
+
+        key_schema_dir_name = f"key_schema_{key_schema_id}"
+        self._dump_key_schema(key_schema_id, key_schema_dir_name)
 
         pickle.dump(serializable_message, self.file)
 
-    def _dump_schemata(self, key_schema_id, value_schema_id):
-        directory = self.directory / self.schema_dir_name
-        directory.mkdir()
+    def _dump_key_schema(self, key_schema_id, schema_dir_name: str):
+        directory = self.directory / schema_dir_name
+        if directory.exists():
+            return
+        directory.mkdir(exist_ok=True)
         (directory / "key_schema.avsc").write_text(
             json.dumps(self.schema_registry_client.get_schema_from_id(key_schema_id).original_schema), encoding="utf-8"
         )
+
+    def _dump_value_schema(self, value_schema_id, schema_dir_name: str):
+        directory = self.directory / schema_dir_name
+        if directory.exists():
+            return
+        directory.mkdir(exist_ok=True)
         (directory / "value_schema.avsc").write_text(
             json.dumps(self.schema_registry_client.get_schema_from_id(value_schema_id).original_schema),
             encoding="utf-8",
         )
-
-    def schema_changed(self, decoded_message: DecodedAvroMessage) -> bool:
-        return (
-            self.current_value_schema_id != decoded_message.value_schema_id and decoded_message.value is not None
-        ) or self.current_key_schema_id != decoded_message.key_schema_id
 
 
 class StdOutAvroWriter(StdOutWriter):
@@ -86,10 +90,13 @@ class AvroFileReader(FileReader):
                 record = pickle.load(self.file)
             except EOFError:
                 return
-            schema_directory = self.directory / record["schema_directory_name"]
+            key_schema_directory = self.directory / record["key_schema_directory_name"]
+            key_schema = load_schema((key_schema_directory / "key_schema.avsc").read_text(encoding="utf-8"))
 
-            key_schema = load_schema((schema_directory / "key_schema.avsc").read_text(encoding="utf-8"))
-            value_schema = load_schema((schema_directory / "value_schema.avsc").read_text(encoding="utf-8"))
+            value_schema = None
+            if record["value_schema_directory_name"] is not None:
+                value_schema_directory = self.directory / record["value_schema_directory_name"]
+                value_schema = load_schema((value_schema_directory / "value_schema.avsc").read_text(encoding="utf-8"))
 
             yield KafkaMessage(
                 json.dumps(record["key"]), json.dumps(record["value"]), record["partition"], key_schema, value_schema
@@ -113,6 +120,7 @@ class AvroMessageDecoder:
     def decode_message_from_avro(self, message: Message):
         key_schema_id, decoded_key = self.decode_bytes(message.key())
         value_schema_id, decoded_value = self.decode_bytes(message.value())
+
         headers = []
         if message.headers():
             for header_key, header_value in message.headers():
@@ -122,13 +130,15 @@ class AvroMessageDecoder:
         decoded_message = DecodedAvroMessage(
             decoded_key, decoded_value, message.partition(), key_schema_id, value_schema_id, headers=headers
         )
-        schema_version = it.count(1)
+
         serializable_message = {
             "key": decoded_message.key,
             "value": decoded_message.value,
             "partition": decoded_message.partition,
-            "schema_directory_name": f"{next(schema_version):04}_{key_schema_id}_{value_schema_id}",
+            "key_schema_directory_name": f"key_schema_{key_schema_id}",
+            "value_schema_directory_name": f"value_schema_{value_schema_id}" if value_schema_id != -1 else None,
         }
+
         return key_schema_id, value_schema_id, decoded_message, serializable_message
